Warnings from Bill Oliver.
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@44098 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/common/classes/fb_string.h
@@ -92,7 +92,7 @@ namespace Firebird
 
 				// Grow buffer exponentially to prevent memory fragmentation
 				if (newSize / 2 < bufferSize)
-					newSize = bufferSize * 2u;
+					newSize = size_t(bufferSize) * 2u;
 
 				// Do not grow buffer beyond string length limit
 				if (newSize > max_length() + 1)
